refactor: restructure project files
